Issue #11 Add "Call" buttons to Thunderbird address book
This allows to call the phone numbers that the user keeps
in his Thunderbird address book, by clicking a button
directly next to the phone number.

This is hooking up both the contact overview in the
main AB window as well as the contact edit dialog.
However, only the latter works at the moment, see comments.